+++ include/workflow/task_bag.hpp
@@ -8,7 +8,7 @@ struct task_bag {
     double const workload;
     double const input_data_size;
     double const output_data_size;
-    size_t const memory_requirement;
+    double const memory_requirement;
     size_t const cardinality;
 };
